requirements.txt
alembic==1.16.4
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.4
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pydantic==2.7.1
cryptography==46.0.4
PyJWT==2.9.0
psycopg2-binary==2.9.9
Pygments==2.19.2
pyotp==2.9.0
pypng==0.20220715.0
pytest==8.4.1
pytest-mock==3.14.1
python-dotenv==1.0.1
qrcode==7.4.2
rich==13.9.4
SQLAlchemy==2.0.32
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
redis
celery
prometheus-client
sentry-sdk
pytest-flask
black
ruff
mypy
bandit
argon2-cffi
email-validator
flask-cors
flask-swagger-ui
flasgger
pytest-cov
tox
coverage
pre-commit
Sphinx
sphinx-rtd-theme
twine
wheel
readme-renderer
docutils
Pygments
requests
pytest-mock
pytest-xdist
Faker
httpx
pytest-sugar
pytest-clarity
